FE/Qt: bugref:9973: Ups..
svn:sync-xref-src-repo-rev: r143449

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMachineView.cpp
@@ -1,4 +1,4 @@
-/* $Id: UIMachineView.cpp 88248 2021-03-22 15:39:12Z sergey.dubov@oracle.com $ */
+/* $Id: UIMachineView.cpp 88249 2021-03-22 15:44:52Z sergey.dubov@oracle.com $ */
 /** @file
  * VBox Qt GUI - UIMachineView class implementation.
  */
@@ -399,8 +399,8 @@ void UIMachineView::sltHandleActionTriggerViewScreenToggle(int iScreen, bool fEn
          * Do not send a hint if nothing has changed to prevent the guest being notified about its own changes. */
         if (   !isFullscreenOrSeamless()
             && uisession()->isGuestSupportsGraphics()
-            && (   m_pFrameBuffer->width() != (int)uWidth
-                || m_pFrameBuffer->height() != (int)uHeight
+            && (   m_pFrameBuffer->width() != uWidth
+                || m_pFrameBuffer->height() != uHeight
                 || uisession()->isScreenVisible(screenId()) != uisession()->isScreenVisibleHostDesires(screenId())))
             storeGuestSizeHint(QSize(uWidth, uHeight));
